♻️(api) move API definition files to joanie core app
The API code is related to an app so its place is in the related
app.
