+++ src/backend/joanie/urls.py
@@ -16,7 +16,7 @@ Including another URLconf
 from django.contrib import admin
 from django.urls import path
 
-from . import api
+from joanie.core import api
 
 urlpatterns = [
     path("admin/", admin.site.urls),
